whatpm/t/tree-test-1.dat

#data
<!Doctype html><body>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE XHTML><title></title><body>
#errors
not HTML5
#document
| <!DOCTYPE XHTML>
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML5><title></title><body>
#errors
not HTML5
#document
| <!DOCTYPE HTML5>
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE html4><title></title><body>
#errors
not HTML5
#document
| <!DOCTYPE html4>
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML 2.0//EN"><title></title><body>
#errors
not HTML5 (quirks doctype)
#document
| <!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML 2.0//EN">
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML PUBLIC "HTML"><title></title><body>
#errors
not HTML5 (quirks doctype)
#document
| <!DOCTYPE HTML PUBLIC "HTML">
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN"><title></title><body>
#errors
not HTML5 (quirks doctype)
#document
| <!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN">
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN"
    "http://www.w3.org/TR/html4/loose.dtd"><title></title><body>
#errors
not HTML5 (limited quirks doctype)
#document
| <!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Frameset//EN"><title></title><body>
#errors
not HTML5 (quirks doctype)
#document
| <!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Frameset//EN">
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Frameset//EN"
    "http://www.w3.org/TR/html4/frameset.dtd"><title></title><body>
#errors
not HTML5 (limited quirks doctype)
#document
| <!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Frameset//EN" "http://www.w3.org/TR/html4/frameset.dtd">
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN"><title></title><body>
#errors
not HTML5 (limited quirks doctype)
#document
| <!DOCTYPE HTML PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN">
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN"
    "http://www.w3.org/TR/html4/loose.dtd"><title></title><body>
#errors
not HTML5 (limited quirks doctype)
#document
| <!DOCTYPE HTML PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML PUBLIC "-//W3C//DTD XHTML 1.0 Frameset//EN"><title></title><body>
#errors
not HTML5 (limited quirks doctype)
#document
| <!DOCTYPE HTML PUBLIC "-//W3C//DTD XHTML 1.0 Frameset//EN">
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML PUBLIC "-//W3C//DTD XHTML 1.0 Frameset//EN"
    "http://www.w3.org/TR/html4/loose.dtd"><title></title><body>
#errors
not HTML5 (limited quirks doctype)
#document
| <!DOCTYPE HTML PUBLIC "-//W3C//DTD XHTML 1.0 Frameset//EN" "http://www.w3.org/TR/html4/loose.dtd">
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML PUBLIC "HTML5"
    "http://www.w3.org/TR/html4/loose.dtd"><title></title><body>
#errors
not HTML5 (no quirks doctype)
#document
| <!DOCTYPE HTML PUBLIC "HTML5" "http://www.w3.org/TR/html4/loose.dtd">
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML SYSTEM "http://www.ibm.com/data/dtd/v11/ibmxhtml1-transitional.dtd"><title></title><body>
#errors
not HTML5 (quirks doctype)
#document
| <!DOCTYPE HTML SYSTEM "http://www.ibm.com/data/dtd/v11/ibmxhtml1-transitional.dtd">
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML><!DOCTYPE HTML><title></title><body>
#errors
duplicate DOCTYPE
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML><!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN"><title></title><body>
#errors
duplicate DOCTYPE
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN"><!DOCTYPE HTML><title></title><body>
#errors
not HTML5
duplicate DOCTYPE
#document
| <!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN">
| <html>
|   <head>
|     <title>
|   <body>

#data
a<html>
#errors
no doctype
<html> in body
#document
| <html>
|   <head>
|   <body>
|     "a"

#data
a<html a>
#errors
no doctype
<html> in body
#document
| <html>
|   a=""
|   <head>
|   <body>
|     "a"

#data
<p>x<html a>y
#errors
no doctype
<html> in body
#document
| <html>
|   a=""
|   <head>
|   <body>
|     <p>
|       "xy"

#data
<!DOCTYPE HTML><html> <head><body>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML><html> 
<head><body>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML><html>   <!---->   <head><body>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <!--  -->
|   <head>
|   <body>

#data
<!DOCTYPE HTML><html>  <!DOCTYPE HTML>  <head><body>
#errors
<!DOCTYPE> in <html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML><html a>  <html b>  <head><body>
#errors
second <html>
#document
| <!DOCTYPE HTML>
| <html>
|   a=""
|   b=""
|   <head>
|   <body>

#data
<!DOCTYPE HTML><html>   </end>  <head><body>
#errors
unmatched </end>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML><html title="t"><head><title></title></head><body></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   title="t"
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML><html manifest=manifest.dat><title></title><body>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   manifest="manifest.dat"
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML><html><head title="t"><title></title></head><body></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     title="t"
|     <title>
|   <body>

#data
<!DOCTYPE HTML><html><head><title class="t"></title></head><body></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|       class="t"
|   <body>

#data
<!DOCTYPE HTML><html><head></head><body><title class="a"></title></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <title>
|       class="a"

#data
<!DOCTYPE HTML><html><head></head><body></body><title class="a"></title></html>
#errors
data after body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <title>
|       class="a"

#data
<!DOCTYPE HTML><html><head><title></title></head><body title="t"></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     title="t"

#data
<!DOCTYPE HTML><html><head><title></title><meta name="a"></head><body></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|     <meta>
|       name="a"
|   <body>

#data
<!DOCTYPE HTML><html><head><title></title></head><body><meta name="a"></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     <meta>
|       name="a"

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body><meta name="a"></html>
#errors
data after body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     <meta>
|       name="a"

#data
<!DOCTYPE HTML><html><head><title></title><link rel="a"></head><body></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|     <link>
|       rel="a"
|   <body>

#data
<!DOCTYPE HTML><html><head><title></title></head><body><link rel="a"></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     <link>
|       rel="a"

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body><link rel="a"></html>
#errors
data after body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     <link>
|       rel="a"

#data
<!DOCTYPE HTML><html><head><title></title><base href="http://"></head><body></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|     <base>
|       href="http://"
|   <body>

#data
<!DOCTYPE HTML><html><head><title></title></head><body><base href="http://"></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     <base>
|       href="http://"

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body><base href="http://"></html>
#errors
data after body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     <base>
|       href="http://"

#data
<!DOCTYPE HTML><p><b><p>X
#errors
<b> is not closed before second <p>
<b> is not closed before end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|       <b>
|     <p>
|       <b>
|         "X"

#data
<!DOCTYPE HTML><b><p>
#errors
<b> not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <b>
|       <p>

#data
<!DOCTYPE HTML><b><p>X
#errors
<b> not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <b>
|       <p>
|         "X"

#data
<!DOCTYPE HTML><ins><p>XX</ins>YYY
#errors
1,32,not closed:p
1,34,not closed:ins
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ins>
|       <p>
|         "XXYYY"

## when script is disabled
#data
<!DOCTYPE HTML><noscript>aaa<noscript>bbb</noscript>ccc</noscript>
#errors
in noscript:#character
unmacthed noscript end tag
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|   <body>
|     "aaa"
|     <noscript>
|       "bbb"
|     "ccc"

## when script is disabled
#data
<!DOCTYPE HTML><body><noscript>aaa<noscript>bbb</noscript>ccc</noscript>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <noscript>
|       "aaa"
|       <noscript>
|         "bbb"
|       "ccc"

## when script is disabled
#data
<!DOCTYPE HTML><head><noscript><noscript>
#errors
<noscript> in noscript
no </noscript>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|   <body>

## when script is disabled
#data
<!DOCTYPE HTML><head><noscript><noscript> 
#errors
<noscript> in noscript
no </noscript>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|       " "
|   <body>

## when script is disabled
#data
<!DOCTYPE HTML><head><noscript> <noscript> 
#errors
<noscript> in noscript
no </noscript>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|       "  "
|   <body>

## When scripting is disabled
#data
<!DOCTYPE HTML><head><noscript><base><title></title><body>
#errors
<base> in <noscript>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|     <base>
|     <title>
|   <body>

## When scripting is disabled
#data
<!DOCTYPE HTML><head><noscript><title></title><body>
#errors
<title> in <noscript> in <head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|     <title>
|   <body>

## When scripting is disabled
#data
<!DOCTYPE HTML><head></head><noscript></noscript>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <noscript>

## When scripting is disabled
#data
<!DOCTYPE HTML><head><noscript><script></script>
#errors
<script> in <noscript> in <head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|     <script>
|   <body>

## When scripting is disabled
#data
<!DOCTYPE HTML><head><noscript><body>
#errors
<body> in <noscript> in <head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|   <body>

## When scripting is disabled
#data
<!DOCTYPE HTML><head><noscript><frameset></frameset>
#errors
<frameset> in <noscript> in <head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|   <frameset>

## When scripting is disabled
#data
<!DOCTYPE HTML><head><noscript><br>
#errors
<br> in <noscript> in <head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|   <body>
|     <br>

## When scripting is disabled
#data
<!DOCTYPE HTML><head><noscript><p>
#errors
<p> in <noscript> in <head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|   <body>
|     <p>

## When scripting is disabled
#data
<!DOCTYPE HTML><head><noscript></p>
#errors
</p> in <noscript> in <head>
unmatched </p>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|   <body>

## When scripting is disabled
#data
<!DOCTYPE HTML><head><noscript></br>
#errors
</br> in <noscript> in <head>
</br> after </head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|   <body>

## When scripting is disabled
#data
<!DOCTYPE HTML><head><noscript></div>
#errors
</div> in <noscript> in <head>
unexpected end of file (no </noscript>)
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|   <body>

## When scripting is disabled
#data
<!DOCTYPE HTML><head><noscript></img>
#errors
</img> in <noscript> in <head>
unexpected end of file (no </noscript>)
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|   <body>

## When scripting is disabled
#data
<!DOCTYPE HTML><head><noscript></head><link><body>
#errors
</head> in <noscript> in <head>
<link> after <head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|     <link>
|   <body>

#data
<!DOCTYPE HTML><head></head></head><link><body>
#errors
</head> after </head>
<link> after </head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <link>
|   <body>

#data
<!DOCTYPE HTML></noscript><head><body>
#errors
</noscript> before <head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML><head></noscript><link><body>
#errors
</noscript> in <head>
<link> in <body>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <link>

#data
<!DOCTYPE HTML><head></head></noscript><link><body>
#errors
</noscript> after </head>
<link> after </head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <link>
|   <body>

## When scripting is disabled
#data
<!DOCTYPE HTML><head><noscript></body><meta></noscript><link><body>
#errors
</body> in <noscript> in <head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|       <meta>
|     <link>
|   <body>

## When scripting is disabled
#data
<!DOCTYPE HTML><head><noscript></html><meta></noscript><link><body>
#errors
</html> in <noscript> in <head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|       <meta>
|     <link>
|   <body>

#data
<!DOCTYPE HTML></p>
#errors
</p> before <head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML></br>
#errors
</br> after </head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML></head></p>
#errors
</p> after </head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML><head></body><link>x
#errors
</body> in head
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <link>
|   <body>
|     "x"

#data
<head></html><link>
#errors
no <!DOCTYPE>
</html> in <head>
#document
| <html>
|   <head>
|     <link>
|   <body>

#data
<head></p><link>
#errors
no <!DOCTYPE>
</p> in head
#document
| <html>
|   <head>
|     <link>
|   <body>

#data
<head></head></p>
#errors
no <!DOCTYPE>
</p> after </head>
#document
| <html>
|   <head>
|   <body>

#data
</p>
#errors
no <!DOCTYPE>
</p> before </head>
#document
| <html>
|   <head>
|   <body>

#data
<head></head></div>
#errors
no <!DOCTYPE>
</div> after </head>
#document
| <html>
|   <head>
|   <body>

#data
<head></head></unknown>
#errors
no <!DOCTYPE>
</unknown> after </head>
#document
| <html>
|   <head>
|   <body>

#data
<head></head></head>
#errors
no <!DOCTYPE>
</head> after </head>
#document
| <html>
|   <head>
|   <body>

#data
<head></head></html>x
#errors
no <!DOCTYPE>
</html> after </head>
#document
| <html>
|   <head>
|   <body>
|     "x"

#data
<head></head></body><body>y</body>
#errors
no <!DOCTYPE>
</body> after </head>
#document
| <html>
|   <head>
|   <body>
|     "y"

#data
<head></head></noscript>
#errors
no <!DOCTYPE>
</noscript> after </head>
#document
| <html>
|   <head>
|   <body>

#data
<head></head></style>
#errors
no <!DOCTYPE>
</style> after </head>
#document
| <html>
|   <head>
|   <body>

#data
<head></br><link>
#errors
no <!DOCTYPE>
</br> in </head>
<link> in </head>
#document
| <html>
|   <head>
|     <link>
|   <body>

#data
<!DOCTYPE HTML><head></head></img>
#errors
</img>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML><head></head></div>
#errors
unmatched </div>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML><head></head></body>
#errors
</body> after </head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML><head></head></html>
#errors
</html> after </head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML><head><title></title></head><base><body>
#errors
<base> after </head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|     <base>
|   <body>

#data
<!DOCTYPE HTML><head><title></title></head><meta><body>
#errors
<meta> after </head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|     <meta>
|   <body>

#data
<!DOCTYPE HTML><head></head><title></title><body>
#errors
<title> after </head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML><iframe>
#errors
no </iframe>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <iframe>

#data
<!DOCTYPE HTML><iframe></iframe>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <iframe>

#data
<!DOCTYPE HTML><iframe>a</iframe>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <iframe>
|       "a"

#data
<!DOCTYPE HTML><noframes>
#errors
no </noframes>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <noframes>

#data
<!DOCTYPE HTML><noframes></noframes>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <noframes>

#data
<!DOCTYPE HTML><noframes>a</noframes>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <noframes>
|       "a"

#data
<!DOCTYPE HTML><noframes>
#errors
no </noframes>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <noframes>

#data
<!DOCTYPE HTML><noframes></noframes>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <noframes>

#data
<!DOCTYPE HTML><noframes>a</noframes>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <noframes>
|       "a"

#data
<!DOCTYPE HTML><xmp>
#errors
no </xmp>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <xmp>

#data
<!DOCTYPE HTML><xmp></xmp>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <xmp>

#data
<!DOCTYPE HTML><xmp>a</xmp>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <xmp>
|       "a"

#data
<!DOCTYPE HTML><html><head><title></title><style></style></head><body>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|     <style>
|   <body>

#data
<!DOCTYPE HTML><html><head><title></title><style type="text/css"></style></head><body>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|     <style>
|       type="text/css"
|   <body>

#data
<!DOCTYPE HTML><html><head><title></title><style type="text/css">p {}</style></head><body>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|     <style>
|       type="text/css"
|       "p {}"
|   <body>

#data
<!DOCTYPE HTML><html><style type="text/css"></style><title></title></head><body>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <style>
|       type="text/css"
|     <title>
|   <body>

#data
<!DOCTYPE HTML><style type="text/css"></style><title></title><body>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <style>
|       type="text/css"
|     <title>
|   <body>

#data
<!DOCTYPE HTML><title></title><body><p>a<style type="text/css"></style>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     <p>
|       "a"
|       <style>
|         type="text/css"

#data
<!DOCTYPE HTML><form><form></form>
#errors
form in form
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>

#data
<!DOCTYPE HTML><ul><li>a<li>c</ul>d
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         "a"
|       <li>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><ul><li>a<div>b<li>c</ul>d
#errors
div closed by li
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         "a"
|         <div>
|           "b"
|       <li>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><dl><dt>a<dt>c</dl>d
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         "a"
|       <dt>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><dl><dt>a<dd>c</dl>d
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         "a"
|       <dd>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><dl><dd>a<dt>c</dl>d
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dd>
|         "a"
|       <dt>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><dl><dd>a<dd>c</dl>d
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dd>
|         "a"
|       <dd>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><dl><dt>a<div>b<dt>c</dl>d
#errors
div closed by dt
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         "a"
|         <div>
|           "b"
|       <dt>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><dl><dt>a<div>b<dd>c</dl>d
#errors
div closed by dd
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         "a"
|         <div>
|           "b"
|       <dd>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><dl><dd>a<div>b<dt>c</dl>d
#errors
div closed by dt
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dd>
|         "a"
|         <div>
|           "b"
|       <dt>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><dl><dd>a<div>b<dd>c</dl>d
#errors
div closed by dd
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dd>
|         "a"
|         <div>
|           "b"
|       <dd>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><body><form><div>a</form>b
#errors
no </div> before </form>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <div>
|         "a"
|     "b"

#data
<!DOCTYPE HTML><body><form><div><div><div>a</form>b
#errors
no </div> before </form>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <div>
|         <div>
|           <div>
|             "a"
|     "b"

#data
<!DOCTYPE HTML><form><p></form>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <p>

#data
<!DOCTYPE HTML><form><p></form>x
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <p>
|     "x"

#data
<!DOCTYPE HTML><form><li></form>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <li>

#data
<!DOCTYPE HTML><form><li></form>y
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <li>
|     "y"

#data
<!DOCTYPE HTML><form><dt></form>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <dt>

#data
<!DOCTYPE HTML><form><dt></form>x
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <dt>
|     "x"

#data
<!DOCTYPE HTML><form><dd></form>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <dd>

#data
<!DOCTYPE HTML><form><dd></form>x
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <dd>
|     "x"

#data
<!DOCTYPE HTML><ul></body>
#errors
ul not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>

#data
<!DOCTYPE HTML><ul><li></body>
#errors
ul not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>

#data
<!DOCTYPE HTML><dl></body>
#errors
dl not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>

#data
<!DOCTYPE HTML><dl><dt></body>
#errors
dl not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>

#data
<!DOCTYPE HTML><dl><dd></body>
#errors
dl not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dd>

#data
<!DOCTYPE HTML><p></body>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>

#data
<!DOCTYPE HTML><table>  </table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       "  "

#data
<!DOCTYPE HTML><table>  a</table>
#errors
char in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     <table>
|       "  "

#data
<!DOCTYPE HTML><table>  a b</table>
#errors
char in <table>
x
3
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a b"
|     <table>
|       "  "

#data
<!DOCTYPE HTML><table>a b </table>
#errors
char in <table>
x
x
4
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a b "
|     <table>

#data
<!DOCTYPE HTML><table><br> </table>
#errors
<br> in <table>
char in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <br>
|     " "
|     <table>

#data
<!DOCTYPE HTML><table><tbody>a </table>
#errors
char in <tbody>
x2
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a "
|     <table>
|       <tbody>

#data
<!DOCTYPE HTML><table><tr>a </table>
#errors
char in <tr>
x2
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a "
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table>a<tr> </table>
#errors
char in <table>
char in <tr>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a "
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody><br> </table>
#errors
<br> in <tbody>
char in <tbody>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <br>
|     " "
|     <table>
|       <tbody>

#data
<!DOCTYPE HTML><table><tr><br> </table>
#errors
<br> in <tr>
char in <tr>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <br>
|     " "
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table a><table b>  </table></table>
#errors
<table> in <table>
unmatched </table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       a=""
|     <table>
|       b=""
|       "  "

#data
<!DOCTYPE HTML><table a><table b>  </table> </table>
#errors
<table> in <table>
unmatched </table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       a=""
|     <table>
|       b=""
|       "  "
|     " "

#data
<!DOCTYPE HTML><table><tr><td><table>  </table></td>x</table>
#errors
char in <tr>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "x"
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <table>
|               "  "

#data
<!DOCTYPE HTML><table><tr><td></td>x<table>  </table></table>
#errors
char in <tr>
no </table> before <table>
unmatched </table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "x"
|     <table>
|       <tbody>
|         <tr>
|           <td>
|     <table>
|       "  "

#data
<!DOCTYPE HTML><table><tr><td><table>x</table><table> </table></table>
#errors
char in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "x"
|             <table>
|             <table>
|               " "

#data
<!DOCTYPE HTML><table><title>a</title></table>
#errors
<title> in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <title>
|       "a"
|     <table>

#data
<!DOCTYPE HTML><table><li><table></table>
#errors
<li> in <table>
<table> in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <li>
|     <table>
|     <table>

#data
<!DOCTYPE HTML><table><input> </table>
#errors
<input> in <table>
char in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <input>
|     " "
|     <table>

#data
<!DOCTYPE HTML><table><input type="text"> </table>
#errors
<input> in <table>
char in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <input>
|       type="text"
|     " "
|     <table>

#data
<!DOCTYPE HTML><table><input type=hidden> </table>
#errors
<input type=hidden> in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <input>
|         type="hidden"
|       " "

#data
<!DOCTYPE HTML><table> <input> </table>
#errors
<input> in <table>
char in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <input>
|     " "
|     <table>
|       " "

#data
<!DOCTYPE HTML><table>a<input> </table>
#errors
char in <table>
<input> in <table>
char in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     <input>
|     " "
|     <table>

#data
<!DOCTYPE HTML><table> <input type="text"> </table>
#errors
<input> in <table>
char in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <input>
|       type="text"
|     " "
|     <table>
|       " "

#data
<!DOCTYPE HTML><table>a<input type=text> </table>
#errors
char in <table>
<input> in <table>
char in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     <input>
|       type="text"
|     " "
|     <table>

#data
<!DOCTYPE HTML><table> <input type=hidden> </table>
#errors
<input type=hidden> in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       " "
|       <input>
|         type="hidden"
|       " "

#data
<!DOCTYPE HTML><table>a<input type=hidden> </table>
#errors
char in <table>
<input> in <table>
char in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     <input>
|       type="hidden"
|     " "
|     <table>

#data
<!DOCTYPE HTML><table><input type="HiddeN"> </table>
#errors
<input type=hidden> in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <input>
|         type="HiddeN"
|       " "

#data
<!DOCTYPE HTML><table><thead><input type=hidden></table>
#errors
<input type=hidden> in <thead>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <thead>
|         <input>
|           type="hidden"

#data
<!DOCTYPE HTML><table><thead>a<input type=hidden></table>
#errors
char in <thead>
<input type=hidden> in <thead>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     <input>
|       type="hidden"
|     <table>
|       <thead>

#data
<!DOCTYPE HTML><table><tr><input type=hidden></table>
#errors
<input type=hidden> in <tr>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <input>
|             type="hidden"

#data
<!DOCTYPE HTML><table><tr>a<input type=hidden></table>
#errors
char in <tr>
<input type=hidden> in <tr>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     <input>
|       type="hidden"
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><script></script></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <script>

#data
<!DOCTYPE HTML><table> <script></script> </table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       " "
|       <script>
|       " "

#data
<!DOCTYPE HTML><table>a<script></script> </table>
#errors
char in <table>
<script> in <table>
char in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     <script>
|     " "
|     <table>

#data
<!DOCTYPE HTML><table><tfoot><script></script> </table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tfoot>
|         <script>
|         " "

#data
<!DOCTYPE HTML><table><tfoot>a<script></script> </table>
#errors
char in <tfoot>
<script> in <tfoot>
char in <tfoot>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     <script>
|     " "
|     <table>
|       <tfoot>

#data
<!DOCTYPE HTML><table><tr> <script></script> </table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           " "
|           <script>
|           " "

#data
<!DOCTYPE HTML><table><tr>a<script></script> </table>
#errors
char in <tr>
<script> in <tr>
char in <tr>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     <script>
|     " "
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><style></style></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <style>

#data
<!DOCTYPE HTML><table> <style></style> </table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       " "
|       <style>
|       " "

#data
<!DOCTYPE HTML><table>a<style></style> </table>
#errors
char in <table>
<style> in <table>
char in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     <style>
|     " "
|     <table>

#data
<!DOCTYPE HTML><table><tfoot><style></style> </table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tfoot>
|         <style>
|         " "

#data
<!DOCTYPE HTML><table><tfoot>a<style></style> </table>
#errors
char in <tfoot>
<style> in <tfoot>
char in <tfoot>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     <style>
|     " "
|     <table>
|       <tfoot>

#data
<!DOCTYPE HTML><table><tr> <style></style> </table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           " "
|           <style>
|           " "

#data
<!DOCTYPE HTML><table><tr>a<style></style> </table>
#errors
char in <tr>
<style> in <tr>
char in <tr>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     <style>
|     " "
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><b>x<i>y<div>z</b>a</i>b
#errors
1,26,in table:b
1,27,in table:#character
1,30,in table:i
1,31,in table:#character
1,36,in table:div
1,37,in table:#character
1,41,in table:/b
1,41,not closed:div
1,42,in table:#character
1,46,in table:/i
1,46,not closed:div
1,46,in table:#character
no </div></table> at the end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <b>
|       "x"
|       <i>
|         "y"
|     <i>
|     <div>
|       <i>
|         <b>
|           "z"
|         "a"
|       "b"
|     <table>

#data
<!DOCTYPE HTML><table><tr><td></body>
#errors
unmatched </body>
table not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>

#data
<!DOCTYPE HTML><table><tr><th></body>
#errors
unmatched </body>
table not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <th>

#data
<!DOCTYPE HTML><table><thead><tr><td></body>
#errors
unmatched </body>
table not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <thead>
|         <tr>
|           <td>

#data
<!DOCTYPE HTML><isindex>
#errors
isindex
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <hr>
|       <p>
|         <label>
|           "This is a searchable index. Insert your search keywords here: "
|           <input>
|             name="isindex"
|       <hr>

#data
<!DOCTYPE HTML><isindex action="http://test/">
#errors
isindex
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       action="http://test/"
|       <hr>
|       <p>
|         <label>
|           "This is a searchable index. Insert your search keywords here: "
|           <input>
|             name="isindex"
|       <hr>

#data
<!DOCTYPE HTML><isindex prompt="Input:">
#errors
isindex
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <hr>
|       <p>
|         <label>
|           "Input:"
|           <input>
|             name="isindex"
|       <hr>

#data
<!DOCTYPE HTML><isindex name="textarea">
#errors
isindex
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <hr>
|       <p>
|         <label>
|           "This is a searchable index. Insert your search keywords here: "
|           <input>
|             name="isindex"
|       <hr>

#data
<!DOCTYPE HTML><isindex type="checkbox">
#errors
isindex
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <hr>
|       <p>
|         <label>
|           "This is a searchable index. Insert your search keywords here: "
|           <input>
|             name="isindex"
|             type="checkbox"
|       <hr>

#data
<!DOCTYPE HTML><form><isindex></form>
#errors
isindex
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>

#data
<!DOCTYPE HTML><h1><a><h2>
#errors
unexpected end of file (no </h1></a></h2>)
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <h1>
|       <a>
|         <h2>

#data
<!DOCTYPE HTML><h1><p></h1>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <h1>
|       <p>

#data
<!DOCTYPE HTML><h1><li></h1>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <h1>
|       <li>

#data
<!DOCTYPE HTML><h1><dt></h1>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <h1>
|       <dt>

#data
<!DOCTYPE HTML><h1><dd></h1>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <h1>
|       <dd>

#data
<!DOCTYPE HTML><h1></h1>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <h1>

#data
  <p>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <p>

#data
  a
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     "a"

#data
<!---->  <p>
#errors
no DOCTYPE
#document
| <!--  -->
| <html>
|   <head>
|   <body>
|     <p>

#data
<!DOCTYPE HTML>  <p>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>

#data
<!DOCTYPE HTML><div></div></div>
#errors
unmatched </div>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <div>

#data
<!DOCTYPE HTML><div><span></div>
#errors
unclosed <span>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <div>
|       <span>

#data
<!DOCTYPE HTML><listing>
</listing>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <listing>

#data
<!DOCTYPE HTML><listing>
a</listing>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <listing>
|       "a"

#data
<!DOCTYPE HTML><listing>a
</listing>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <listing>
|       "a
"

#data
<!DOCTYPE HTML><listing>

</listing>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <listing>
|       "
"

#data
<!DOCTYPE HTML><listing> 
</listing>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <listing>
|       " 
"

#data
<!DOCTYPE HTML><pre><!---->
a</pre>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <pre>
|       <!--  -->
|       "
a"

#data
<!DOCTYPE HTML><listing><!---->
a</listing>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <listing>
|       <!--  -->
|       "
a"

#data
<!DOCTYPE HTML><textarea><!---->
a</textarea>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <textarea>
|       "<!---->
a"

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html> 
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     " "

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html>  
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     "  "

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html>a
#errors
char after </body>
char after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     "a"

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html> a
#errors
char after </html>
char after </body>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     " a"

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html>a 
#errors
char after </html>
char after </body>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     "a "

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html>aaa
#errors
char after </html>
char after </body>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     "aaa"

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html><!---->
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
| <!--  -->

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html><!--aa-->
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
| <!-- aa -->

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html><!DOCTYPE HTML>
#errors
<!DOCTYPE> after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html><p>
#errors
start tag after </html>
start tag after </body>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     <p>

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html></p>
#errors
end tag after </html>
end tag after </body>
</p> interpreted as <p></p>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     <p>

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html>  <p>
#errors
start tag after </html>
start tag after </body>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     "  "
|     <p>

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html>aa<p>
#errors
char after </html>
char after </body>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     "aa"
|     <p>

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html> 
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   " "

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html>  
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   "  "

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html>a
#errors
char after </html>
char after </frameset>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html> a
#errors
char after </html>
char after </frameset>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   " "

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html>a 
#errors
char after </html>
char after </frameset>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   " "

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html><!---->
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
| <!--  -->

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html><!--a-->
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
| <!-- a -->

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html> <!---->
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   " "
| <!--  -->

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html>a<!---->
#errors
char after </html>
char after </frameset>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   <!--  -->

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html><!DOCTYPE HTML>
#errors
<!DOCTYPE> after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html> <!DOCTYPE HTML>
#errors
<!DOCTYPE> after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   " "

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html><p>
#errors
start tag after </html>
start tag after </frameset>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html> <p>
#errors
start tag after </html>
start tag after </frameset>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   " "

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html></p>
#errors
end tag after </html>
end tag after </frameset>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html> </p>
#errors
end tag after </html>
end tag after </frameset>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   " "

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html></html>
#errors
</html> after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html>a</html>
#errors
char after </html>
char after </frameset>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html><noframes></noframes>
#errors
<noframes> after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   <noframes>

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html><noframes>
#errors
<noframes> after </html>
no </noframes>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   <noframes>

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html><noframes><p></noframes>
#errors
<noframes> after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   <noframes>
|     "<p>"

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html><noframes></noframes>a
#errors
<noframes> after </html>
char after </frameset>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   <noframes>

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html> <noframes></noframes> 
#errors
<noframes> after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   " "
|   <noframes>
|   " "

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset> </frameset></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|     " "

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body> </html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     " "

#data
<!DOCTYPE HTML><frameset></frameset><noframes>
#errors
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>
|   <noframes>

#data
<!DOCTYPE HTML><frameset></frameset><noframes></noframes>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>
|   <noframes>

#data
<!DOCTYPE HTML><frameset></frameset><noframes>abcdefg
#errors
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>
|   <noframes>
|     "abcdefg"

#data
<!DOCTYPE HTML><frameset></frameset><noframes>abcdefg</noframes>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>
|   <noframes>
|     "abcdefg"

#data
<!DOCTYPE HTML><frameset></frameset><noframes>aa<p>bb</noframes>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>
|   <noframes>
|     "aa<p>bb"

#data
<!DOCTYPE HTML><frameset></frameset><noframes><noframes></noframes>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>
|   <noframes>
|     "<noframes>"

#data
<!DOCTYPE HTML><frameset></frameset><noframes></noframe></noframes>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>
|   <noframes>
|     "</noframe>"

#data
</body></html><html a>
#errors
no <!DOCTYPE>
</body> before head
</html> before head
<html> in body
#document
| <html>
|   a=""
|   <head>
|   <body>

#data
</html>
#errors
no <!DOCTYPE>
</html> before head
#document
| <html>
|   <head>
|   <body>

#data
<!----></html>x
#errors
no <!DOCTYPE>
</html> before head
#document
| <!--  -->
| <html>
|   <head>
|   <body>
|     "x"

#data
<body></body></html><html a>
#errors
no <!DOCTYPE>
<html> after </html>
<html> is not the first start tag
#document
| <html>
|   a=""
|   <head>
|   <body>

#data
<frameset></frameset></html><html a>
#errors
no <!DOCTYPE>
<html> after </html>
<html> is not the first start tag
#document
| <html>
|   a=""
|   <head>
|   <frameset>

#data
<!DOCTYPE HTML><html><head></head><frameset></frameset></html><html a>
#errors
<html> after </html>
<html> is not the first start tag
#document
| <!DOCTYPE HTML>
| <html>
|   a=""
|   <head>
|   <frameset>

#data
<frameset></frameset>
#errors
no <!DOCTYPE>
#document
| <html>
|   <head>
|   <frameset>

#data
<frameset></frameset></frameset>
#errors
no <!DOCTYPE>
unmatched </frameset>
#document
| <html>
|   <head>
|   <frameset>

#data
<frameset></frameset></frameset></frameset>
#errors
no <!DOCTYPE>
unmatched </frameset>
unmatched </frameset>
#document
| <html>
|   <head>
|   <frameset>

#data
<!DOCTYPE HTML><table><caption>test</caption>a</table>
#errors
a in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     <table>
|       <caption>
|         "test"

#data
<!DOCTYPE HTML><table><caption>test</table>a
#errors
no </caption>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "test"
|     "a"

#data
<!DOCTYPE HTML><table><caption>test</body>a
#errors
unmatched </body>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "testa"

#data
<!DOCTYPE HTML><table><caption>test</col>a
#errors
unmatched </col>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "testa"

#data
<!DOCTYPE HTML><table><caption>test</colgroup>a
#errors
unmatched </colgroup>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "testa"

#data
<!DOCTYPE HTML><table><caption>test</html>a
#errors
unmatched </html>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "testa"

#data
<!DOCTYPE HTML><table><caption>test</html> 
#errors
unmatched </html>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "test "

#data
<!DOCTYPE HTML><table><caption>test</tbody>a
#errors
unmatched </tbody>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "testa"

#data
<!DOCTYPE HTML><table><caption>test</thead>a
#errors
unmatched </thead>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "testa"

#data
<!DOCTYPE HTML><table><caption>test</tfoot>a
#errors
unmatched </tfoot>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "testa"

#data
<!DOCTYPE HTML><table><caption>test</tr>a
#errors
unmatched </tr>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "testa"

#data
<!DOCTYPE HTML><table><caption>test</td>a
#errors
unmatched </td>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "testa"

#data
<!DOCTYPE HTML><table><caption>test</th>a
#errors
unmatched </th>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "testa"

#data
<!DOCTYPE HTML><table><caption>a<caption>b
#errors
no </caption>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "a"
|       <caption>
|         "b"

#data
<!DOCTYPE HTML><table><colgroup>  </table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <colgroup>
|         "  "

#data
<!DOCTYPE HTML><table><colgroup></colgroup></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <colgroup>

#data
<!DOCTYPE HTML><table><caption>a<col>b
#errors
no </caption>
"b" in table
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "b"
|     <table>
|       <caption>
|         "a"
|       <colgroup>
|         <col>

#data
<!DOCTYPE HTML><table><caption>a<colgroup>b
#errors
no </caption>
"b" in colgroup
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "b"
|     <table>
|       <caption>
|         "a"
|       <colgroup>

#data
<!DOCTYPE HTML><table><caption>a<thead>b
#errors
no </caption>
"b" in thead
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "b"
|     <table>
|       <caption>
|         "a"
|       <thead>

#data
<!DOCTYPE HTML><table><caption>a<tbody>b
#errors
no </caption>
"b" in tbody
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "b"
|     <table>
|       <caption>
|         "a"
|       <tbody>

#data
<!DOCTYPE HTML><table><caption>a<tfoot>b
#errors
no </caption>
"b" in tfoot
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "b"
|     <table>
|       <caption>
|         "a"
|       <tfoot>

#data
<!DOCTYPE HTML><table><caption>a<tr>b
#errors
no </caption>
"b" in tr
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "b"
|     <table>
|       <caption>
|         "a"
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><caption>a<td>b
#errors
no </caption>
no <tr>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "a"
|       <tbody>
|         <tr>
|           <td>
|             "b"

#data
<!DOCTYPE HTML><table><caption>a<th>b
#errors
no </caption>
no <tr>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "a"
|       <tbody>
|         <tr>
|           <th>
|             "b"

#data
<!DOCTYPE HTML>a<caption>b
#errors
<caption> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML><table><caption><span><tbody>
#errors
<caption> not closed
<span> not closed
<table> not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <span>
|       <tbody>

#data
<!DOCTYPE HTML><table><caption><li><tbody>
#errors
<caption> not closed
<table> not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <li>
|       <tbody>

#data
<!DOCTYPE HTML><table><caption><p><tbody>
#errors
<caption> not closed
<table> not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <p>
|       <tbody>

#data
<!DOCTYPE HTML><table><caption><dt><tbody>
#errors
<caption> not closed
<table> not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <dt>
|       <tbody>

#data
<!DOCTYPE HTML><table><caption><dd><tbody>
#errors
<caption> not closed
<table> not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <dd>
|       <tbody>

#data
<!DOCTYPE HTML>a<col>b
#errors
<col> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<colgroup>b
#errors
<colgroup> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<frame>b
#errors
<frame> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<frameset>b
#errors
<frameset> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<frameset><frame>b
#errors
<frameset> in body
<frame> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<head>b
#errors
<head> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<head><link>b
#errors
<head> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     <link>
|     "b"

#data
<!DOCTYPE HTML>a<option>b
#errors
<option> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<optgroup>b
#errors
<optgroup> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<tbody>b
#errors
<tbody> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<tbody><tr>b
#errors
<tbody> in body
<tr> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<thead>b
#errors
<thead> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<tfoot>b
#errors
<tfoot> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<tr>b
#errors
<tr> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<td>b
#errors
<td> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<th>b
#errors
<th> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML><table><tr><td>a<tr><td>b</table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "a"
|         <tr>
|           <td>
|             "b"

#data
<!DOCTYPE HTML><table><tr><td>a<thead><tr><td>b</table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "a"
|       <thead>
|         <tr>
|           <td>
|             "b"

#data
<!DOCTYPE HTML><table><tr><td>a<tbody><tr><td>b</table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "a"
|       <tbody>
|         <tr>
|           <td>
|             "b"

#data
<!DOCTYPE HTML><table><tr><td>a<tfoot><tr><td>b</table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "a"
|       <tfoot>
|         <tr>
|           <td>
|             "b"

#data
<!DOCTYPE HTML><table><tr><td>a<caption>b</caption></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "a"
|       <caption>
|         "b"

#data
<!DOCTYPE HTML><table><tr><td>a</th></table>
#errors
unmatched </th>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "a"

#data
<!DOCTYPE HTML><table><tr><td>a</th>b</table>
#errors
unmatched </th>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "ab"

#data
<!DOCTYPE HTML><table><tr><td>a</body>b
#errors
unmatched </body>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "ab"

#data
<!DOCTYPE HTML><table><tr><td>a</html>b
#errors
unmatched </html>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "ab"

#data
<!DOCTYPE HTML><table><tr><td>a</caption>b
#errors
unmatched </caption>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "ab"

#data
<!DOCTYPE HTML><table><tr><td>a</colgroup>b
#errors
unmatched </colgroup>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "ab"

#data
<!DOCTYPE HTML><table><tr><td>a</col>b
#errors
unmatched </col>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "ab"

#data
<!DOCTYPE HTML><table><tr><td>a</thead>b
#errors
unmatched </thead>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "ab"

#data
<!DOCTYPE HTML><table><tr><td>a</tfoot>b
#errors
unmatched </tfoot>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "ab"

#data
<!DOCTYPE HTML><table><tfoot><tr><td>a</tbody>b
#errors
unmatched </tbody>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tfoot>
|         <tr>
|           <td>
|             "ab"

#data
<!DOCTYPE HTML><table><tr><td>a</tbody>b
#errors
"b" in table
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "b"
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "a"

#data
<!DOCTYPE HTML><table><thead><tr><td>a</thead>b
#errors
"b" in table
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "b"
|     <table>
|       <thead>
|         <tr>
|           <td>
|             "a"

#data
<!DOCTYPE HTML><table><tfoot><tr><td>a</tfoot>b
#errors
"b" in table
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "b"
|     <table>
|       <tfoot>
|         <tr>
|           <td>
|             "a"

#data
<!DOCTYPE HTML><table><tbody><caption>a</caption>
#errors
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|       <caption>
|         "a"

#data
<!DOCTYPE HTML><table><caption><li></caption></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <li>

#data
<!DOCTYPE HTML><table><caption><p></caption></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <p>

#data
<!DOCTYPE HTML><table><caption><dt></caption></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <dt>

#data
<!DOCTYPE HTML><table><caption><dd></caption></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <dd>

#data
<!DOCTYPE HTML><table><caption><p></table>
#errors
<caption> not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <p>

#data
<!DOCTYPE HTML><table><caption><li></table>
#errors
<caption> not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <li>

#data
<!DOCTYPE HTML><table><caption><dd></table>
#errors
<caption> not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <dd>

#data
<!DOCTYPE HTML><table><caption><dt></table>
#errors
<caption> not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <dt>

#data
<!DOCTYPE HTML><table><caption><span></table>
#errors
<caption> not closed
<span> not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <span>

#data
<!DOCTYPE HTML><table><tbody><thead><tr></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|       <thead>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody><tfoot><tr></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|       <tfoot>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody><tbody><tr></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><thead><tfoot><tr></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <thead>
|       <tfoot>
|         <tr>

#data
<!DOCTYPE HTML><table><thead><tbody><tr></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <thead>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><thead><thead><tr></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <thead>
|       <thead>
|         <tr>

#data
<!DOCTYPE HTML><table a><thead><table b>
#errors
<table> in table row group
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       a=""
|       <thead>
|     <table>
|       b=""

#data
<!DOCTYPE HTML><table a><tbody><table b>
#errors
<table> in table row group
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       a=""
|       <tbody>
|     <table>
|       b=""

#data
<!DOCTYPE HTML><table a><tfoot><table b>
#errors
<table> in table row group
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       a=""
|       <tfoot>
|     <table>
|       b=""

#data
<!DOCTYPE HTML><table a><tr><table b>
#errors
<table> in table row
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       a=""
|       <tbody>
|         <tr>
|     <table>
|       b=""

#data
<!DOCTYPE HTML><table><thead></tbody><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <thead>
|         <tr>

#data
<!DOCTYPE HTML><table><thead></thead><tr>
#errors
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <thead>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><thead></tfoot><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <thead>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody></thead><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody></tbody><tr>
#errors
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody></tfoot><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tfoot></thead><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tfoot>
|         <tr>

#data
<!DOCTYPE HTML><table><tfoot></tbody><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tfoot>
|         <tr>

#data
<!DOCTYPE HTML><table><tfoot></tfoot><tr>
#errors
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tfoot>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody></tr><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody></body><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody></html><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody></td><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody></th><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody></caption><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody></col><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody></colgroup><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tr><form><td></td></form></table>
#errors
<form> in table
form not closed
</form> in table
ubmatched </form>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|     <table>
|       <tbody>
|         <tr>
|           <td>

#data
<!DOCTYPE HTML><table><tr><td></th></table>
#errors
unmatched </th>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>

#data
<!DOCTYPE HTML><table><tr><th></td></table>
#errors
unmatched </td>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <th>

#data
<!DOCTYPE HTML><table><tr><td><dd></td></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <dd>

#data
<!DOCTYPE HTML><table><tr><td><dt></td></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <dt>

#data
<!DOCTYPE HTML><table><tr><td><li></td></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <li>

#data
<!DOCTYPE HTML><table><tr><td><p></td></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <p>

#data
<!DOCTYPE HTML><table><tr><th><p></th></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <th>
|             <p>

#data
<!DOCTYPE HTML><body></h1>
#errors
unmatched </h1>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
<p><font><form> </form>
#errors
1,4,no DOCTYPE
1,16,not closed:font
unexpected end-of-file
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <font>
|     <form>
|       <font>
|         " "

#data
<p><font><form> </form>a
#errors
no DOCTYPE
<font> not closed before <form>
<font> not closed before </form>
unexpected end-of-file (<font> not closed before end-of-file)
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <font>
|     <form>
|       <font>
|         " "
|     <font>
|       "a"

#data
<table a><tr><table b>
#errors
1,10,no DOCTYPE
1,22,not closed:tr
1,22,not closed:table
#document
| <html>
|   <head>
|   <body>
|     <table>
|       a=""
|       <tbody>
|         <tr>
|     <table>
|       b=""

#data
<!DOCTYPE HTML>foo<html x>
#errors
<html> in body
#document
| <!DOCTYPE HTML>
| <html>
|   x=""
|   <head>
|   <body>
|     "foo"

#data
<!DOCTYPE HTML><select><option><option>
#errors
no </select>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       <option>
|       <option>

#data
<!DOCTYPE HTML><select></p>
#errors
</p> in <select>
no </select>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>

#data
<!DOCTYPE HTML><select></br>
#errors
</br> in <select>
no </select>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>

#data
<!DOCTYPE HTML><select></base>
#errors
</base> in <select>
no </select>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>

#data
<!DOCTYPE HTML><select></div>
#errors
</div> in <select>
no </select>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>

#data
<!DOCTYPE HTML><select></option>
#errors
unmatched </option>
no </select>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>

#data
<!DOCTYPE HTML><select><option></option></select>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       <option>

#data
<!DOCTYPE HTML><select><optgroup></option></select>
#errors
</option> in <optgroup>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       <optgroup>

#data
<!DOCTYPE HTML><select></table>aa
#errors
unmatched </table>
no </select>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       "aa"

#data
<!DOCTYPE HTML><select></tbody>aa
#errors
unmatched </tbody>
no </select>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       "aa"

#data
<!DOCTYPE HTML><select></thead>aa
#errors
unmatched </thead>
no </select>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       "aa"

#data
<!DOCTYPE HTML><select></tfoot>aa
#errors
unmatched </tfoot>
no </select>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       "aa"

#data
<!DOCTYPE HTML><select></tr>aa
#errors
unmatched </tr>
no </select>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       "aa"

#data
<!DOCTYPE HTML><select></td>aa
#errors
unmatched </td>
no </select>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       "aa"

#data
<!DOCTYPE HTML><select></th>aa
#errors
unmatched </th>
no </select>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       "aa"

#data
<!DOCTYPE HTML><select></caption>aa
#errors
unmatched </caption>
no </select>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       "aa"

#data
<!DOCTYPE HTML><table><tr><td><select></tbody>a
#errors
no </select>
char in <table>
no </table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>

#data
<!DOCTYPE HTML><table><tr><td><select></table>a
#errors
no </select>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>
|     "a"

#data
<!DOCTYPE HTML><table><tr><td><select></thead>a
#errors
unmatched </thead>
unexpected end of file (no </select>, no </table>)
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>
|               "a"

#data
<!DOCTYPE HTML><table><tr><td><select></tfoot>a
#errors
unmatched </tfoot>
unexpected end of file (no </select>, no </table>)
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>
|               "a"

#data
<!DOCTYPE HTML><table><tfoot><tr><td><select></thead>a
#errors
unmatched </thead>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tfoot>
|         <tr>
|           <td>
|             <select>
|               "a"

#data
<!DOCTYPE HTML><table><tr><td><select></tr>a
#errors
no </select>
char in <tbody>
no </table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>

#data
<!DOCTYPE HTML><table><tr><td><select></td>a
#errors
no </select>
char in <tr>
no </table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>

#data
<!DOCTYPE HTML><table><tr><td><select></th>a
#errors
unmatched </th>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>
|               "a"

#data
<!DOCTYPE HTML><table><select></table>a
#errors
<select> in <table>
no </select> before </table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|     <table>
|     "a"

#data
<!DOCTYPE HTML><table><select><option>b</table>a
#errors
<select> in <table>
no </select> before </table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       <option>
|         "b"
|     <table>
|     "a"

#data
<!DOCTYPE HTML><table><tfoot><select></table>a
#errors
<select> in <tfoot>
no </select> before </table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|     <table>
|       <tfoot>
|     "a"

#data
<!DOCTYPE HTML><table><select><option>b</tfoot>a
#errors
<select> in <table>
unmatched </tfoot>
no </table> at the end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       <option>
|         "ba"
|     <table>

#data
<!DOCTYPE HTML><table><tr><select></table>a
#errors
<select> in <tr>
no </select> before </table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|     <table>
|       <tbody>
|         <tr>
|     "a"

#data
<!DOCTYPE HTML><table><tr><select></tr>a</table>
#errors
<select> in <tr>
no </select> before </tr>
char in <tbody>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|     "a"
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tr><td><select><option>b</table>a
#errors
no </select> before </table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>
|               <option>
|                 "b"
|     "a"

#data
<!DOCTYPE HTML><select><input>x
#errors
no </select> before </input>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|     <input>
|     "x"

#data
<!DOCTYPE HTML><select><option>a<input>b
#errors
no </select> before <input>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       <option>
|         "a"
|     <input>
|     "b"

#data
<!DOCTYPE HTML><table><select><input></table>a
#errors
<select> in <table>
no </select> before <input>
<input> in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|     <input>
|     <table>
|     "a"

#data
<!DOCTYPE HTML><table><select><option>b<input></table>a
#errors
<select> in <table>
no </select> before <input>
<input> in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       <option>
|         "b"
|     <input>
|     <table>
|     "a"

#data
<!DOCTYPE HTML><select><table></table>
#errors
<table> in <select>
</table> in <select>
no </select> at the end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>

#data
<!DOCTYPE HTML><select><option><table></table>x
#errors
<table> in <select>
</table> in <select>
no </select> at the end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       <option>
|         "x"

#data
<!DOCTYPE HTML><table><tr><td><select><table></table>x
#errors
no </select> before <table>
no </table> at the end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>
|             <table>
|             "x"

#data
<!DOCTYPE HTML><select><option><caption>x
#errors
<caption> in <select>
no </select> at the end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       <option>
|         "x"

#data
<!DOCTYPE HTML><table><tr><td><select><caption>x
#errors
no </select> before <caption>
no </caption></table> at the end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>
|       <caption>
|         "x"

#data
<!DOCTYPE HTML><select><tbody>x
#errors
<tbody> in <select>
no </select> at the end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       "x"

#data
<!DOCTYPE HTML><table><tr><td><select><tbody><tr><td>x
#errors
no </select> before <tbody>
no </table> at the end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>
|       <tbody>
|         <tr>
|           <td>
|             "x"

#data
<!DOCTYPE HTML><select><tr>x
#errors
<tr> in <select>
no </select> at the end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       "x"

#data
<!DOCTYPE HTML><table><tr><td><select><tr><td>x
#errors
no </select> before <tr>
no </table> at the end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>
|         <tr>
|           <td>
|             "x"

#data
<!DOCTYPE HTML><select><td>x
#errors
<td> in <select>
no </select> at the end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       "x"

#data
<!DOCTYPE HTML><table><tr><td><select><td>x
#errors
no </select> before <table>
no </table> at the end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>
|           <td>
|             "x"

#data
<!DOCTYPE HTML><p><plaintext>
#errors
unexpected end of file (no </plaintext>)
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|     <plaintext>

#data
<!DOCTYPE HTML><p><h1>
#errors
unexpected end of file (no </h1>)
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|     <h1>

#data
<!DOCTYPE HTML><p><table></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|     <table>

#data
<!DOCTYPE HTML><textarea></textarea>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <textarea>

#data
<!DOCTYPE HTML><textarea>
#errors
no </textarea>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <textarea>

#data
<!DOCTYPE HTML><span></html>
#errors
no </span>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <span>

#data
<!DOCTYPE HTML><unknown><p></unknown>
#errors
unmatched </unknown>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <unknown>
|       <p>

#data
<!DOCTYPE HTML><unknown><p></unknown>xx
#errors
unmatched </unknown>
unexpected end of file (no </unknown>)
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <unknown>
|       <p>
|         "xx"

#data
<!DOCTYPE HTML><p></unknown>xx
#errors
unmatched </unknown>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|       "xx"

#data
<!DOCTYPE HTML><b>a<i>b<applet>c</b>d</i>e
#errors
1,37,unmatched end tag:b
1,42,unmatched end tag:i
no </applet> at the end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <b>
|       "a"
|       <i>
|         "b"
|         <applet>
|           "cde"

#data
<!DOCTYPE HTML><li>a<applet>b</li>c
#errors
1,35,unmatched end tag:li
no </applet> at the end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <li>
|       "a"
|       <applet>
|         "bc"

#data
<!DOCTYPE HTML><dt>a<applet>b</dt>c
#errors
1,35,unmatched end tag:dt
no </applet> at the end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dt>
|       "a"
|       <applet>
|         "bc"

#data
<!DOCTYPE HTML><dd>a<applet>b</dd>c
#errors
1,35,unmatched end tag:dd
no </applet> at the end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dd>
|       "a"
|       <applet>
|         "bc"

#data
<!DOCTYPE HTML><x>a<applet>b</x>c
#errors
1,33,unmatched end tag:x
no </applet> at the end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <x>
|       "a"
|       <applet>
|         "bc"

#data
<!DOCTYPE HTML><p><applet><address>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|       <applet>
|         <address>

#data
<!DOCTYPE HTML><p><applet><h1>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|       <applet>
|         <h1>

#data
<!DOCTYPE HTML><p><applet><pre>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|       <applet>
|         <pre>

#data
<!DOCTYPE HTML><p><applet><listing>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|       <applet>
|         <listing>

#data
<!DOCTYPE HTML><p><applet><div>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|       <applet>
|         <div>

#data
<!DOCTYPE HTML><p><applet><form>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|       <applet>
|         <form>

#data
<!DOCTYPE HTML><p><applet><li>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|       <applet>
|         <li>

#data
<!DOCTYPE HTML><p><applet><dt>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|       <applet>
|         <dt>

#data
<!DOCTYPE HTML><p><applet><dd>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|       <applet>
|         <dd>

#data
<!DOCTYPE HTML><p><applet><plaintext>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|       <applet>
|         <plaintext>

#data
<!DOCTYPE HTML><p><applet><table>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|       <applet>
|         <table>

#data
<!DOCTYPE HTML><p><applet><hr>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|       <applet>
|         <hr>

#data
<!DOCTYPE HTML><nobr><applet><nobr>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <nobr>
|       <applet>
|         <nobr>

#data
<!DOCTYPE HTML><button><applet><button>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <button>
|       <applet>
|         <button>

#data
<!DOCTYPE HTML><address><applet></address>x
#errors
unmatched end tag
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <address>
|       <applet>
|         "x"

#data
<!DOCTYPE HTML><h1><applet></h1>x
#errors
unmatched end tag
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <h1>
|       <applet>
|         "x"

#data
<!DOCTYPE HTML><button><applet></button>x
#errors
unmatched end tag
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <button>
|       <applet>
|         "x"

#data
<!DOCTYPE HTML><object><applet></object>x
#errors
unmatched end tag
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <object>
|       <applet>
|         "x"

#data
<!DOCTYPE HTML><p><applet></p>x
#errors
unmatched end tag </p>
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|       <applet>
|         <p>
|         "x"

#data
<!DOCTYPE HTML><dt><applet></dt>x
#errors
unmatched end tag
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dt>
|       <applet>
|         "x"

#data
<!DOCTYPE HTML><dd><applet></dd>x
#errors
unmatched end tag
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dd>
|       <applet>
|         "x"

#data
<!DOCTYPE HTML><pre><applet></pre>x
#errors
unmatched end tag
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <pre>
|       <applet>
|         "x"

#data
<!DOCTYPE HTML><form><applet></form>x
#errors
unmatched end tag
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <applet>
|         "x"

#data
<!DOCTYPE HTML><applet><applet></applet>x
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <applet>
|       <applet>
|       "x"

#data
#errors
not HTML5
#document
| <html>
|   <head>
|   <body>

#data
<!---->
#errors
not HTML5
#document
| <!--  -->
| <html>
|   <head>
|   <body>

#data
    
#errors
not HTML5
#document
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML><!---->
#errors
#document
| <!DOCTYPE HTML>
| <!--  -->
| <html>
|   <head>
|   <body>

#data
<html>
#errors
not HTML5
#document
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML><html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
a
#errors
not HTML5
#document
| <html>
|   <head>
|   <body>
|     "a"

#data
<!DOCTYPE HTML><html><head>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML><html><head><noscript>
#errors
no </noscript>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|   <body>

#data
<!DOCTYPE HTML><html><head></head>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML><html><head></head><body>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML><html><head></head><body><table>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>

#data
<!DOCTYPE HTML><html><head></head><body><table><span>
#errors
<span> in <table>
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <span>
|     <table>

#data
<!DOCTYPE HTML><html><head></head><body><table><caption>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>

#data
<!DOCTYPE HTML><html><head></head><body><table><colgroup>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <colgroup>

#data
<!DOCTYPE HTML><html><head></head><body><table><thead>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <thead>

#data
<!DOCTYPE HTML><html><head></head><body><table><tr>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><html><head></head><body><table><tr><td>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>

#data
<!DOCTYPE HTML><html><head></head><body><select>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>

#data
<!DOCTYPE HTML><html><head></head><body><select><option>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       <option>

#data
<!DOCTYPE HTML><html><head></head><body><select><optgroup>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       <optgroup>

#data
<!DOCTYPE HTML><html><head></head><body></body>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML><html><head></head><body></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML><html><head></head><frameset>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>

#data
<!DOCTYPE HTML><html><head></head><frameset></frameset>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>

#data
<!DOCTYPE HTML><html><head></head><frameset></frameset></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>

#data
<!DOCTYPE HTML><html><head></head><body><li>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <li>

#data
<!DOCTYPE HTML><html><head></head><body><dt>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dt>

#data
<!DOCTYPE HTML><html><head></head><body><dd>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dd>

#data
<!DOCTYPE HTML><html><head></head><body><p>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>

#data
<!DOCTYPE HTML><html><head></head><body><span>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <span>
